Move setup logic to own hook and move related setup_step type into same file
